+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Mini views layout on actives speaker.
 - Set 1 month to default message expires.
 - User-agent format compliance
+- Update SDK to 5.2.15
 
 ## 5.0.6 - 2023-01-10